moved nearly all functions from the explorer to a graphing manager
