+++ HousePartyTranslator.csproj
@@ -119,7 +119,7 @@
     <Compile Include="Helpers\LanguageHelper.cs" />
     <Compile Include="Helpers\Items.cs" />
     <Compile Include="Helpers\Node.cs" />
-    <Compile Include="Managers\GraphingManager.cs" />
+    <Compile Include="StoryExplorerForm\GraphingManager.cs" />
     <Compile Include="Properties\Settings.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
